Allowances handling and manager view completed

+++ src/Utility/Pair.java
@@ -2,8 +2,8 @@ package Utility;
 
 public class Pair {
 	private String key;
-	private Integer value;
-	public Pair(String key, Integer value)
+	private Double value;
+	public Pair(String key, Double value)
 	{
 		this.key=key;
 		this.value=value;
@@ -14,10 +14,10 @@ public class Pair {
 	public void setKey(String key) {
 		this.key = key;
 	}
-	public Integer getValue() {
+	public Double getValue() {
 		return value;
 	}
-	public void setValue(Integer value) {
+	public void setValue(Double value) {
 		this.value = value;
 	}
